Refactor decompression modules
