+++ tower-http/src/decompression/response/future.rs
@@ -1,6 +1,6 @@
 #![allow(unused_imports)]
 
-use super::{body::BodyInner, DecompressionBody};
+use crate::decompression::{body::BodyInner, DecompressionBody};
 use crate::compression_utils::{AcceptEncoding, WrapBody};
 use crate::content_encoding::SupportedEncodings;
 use futures_util::ready;
