+++ tower-http/src/lib.rs
@@ -253,13 +253,6 @@ pub mod sensitive_headers;
 pub mod decompression;
 
 #[cfg(any(
-feature = "decompression-br",
-feature = "decompression-deflate",
-feature = "decompression-gzip"
-))]
-pub mod decompression_request;
-
-#[cfg(any(
     feature = "compression-br",
     feature = "compression-deflate",
     feature = "compression-gzip",
